refactor(tests): main -> cli + processing
