+++ tests/test_processing.py
@@ -1,10 +1,4 @@
-"""
-Integration and functional tests for the __main__ module.
-
-Integration tests involves interactions between a few things (e.g. file system,
-linter logic and conversion process), whilst functional tests involve running
-the full workflow (e.g. simulating user commands).
-"""
+"""Tests for the processing module."""
 
 import re
 import subprocess
@@ -13,7 +7,6 @@ from pathlib import Path
 
 import pytest
 
-from lintquarto.cli import main
 from lintquarto.processing import (
     gather_qmd_files,
     process_qmd,
@@ -175,126 +168,3 @@ def test_validate_no_commas():
     """Unit Test: raises ValueError when path contains a comma."""
     with pytest.raises(ValueError, match="contains a comma"):
         validate_no_commas(["file1.qmd,dir2"], "paths")
-
-
-# =============================================================================
-# 4. __main__()
-# =============================================================================
-
-
-def test_main_runs_functional(tmp_path):
-    """Functional Test: main() runs as a CLI entry point on real .qmd file."""
-    # Create a minimal .qmd file for linting
-    qmd_file = tmp_path / "test.qmd"
-    qmd_file.write_text("# Test\n``````")
-
-    # Run the CLI tool as a subprocess, mimicking user command-line usage and
-    # assert that the process exits with a valid code
-    result = subprocess.run(
-        [
-            sys.executable,
-            "-m",
-            "lintquarto",
-            "-l",
-            CORE_LINTER,
-            "-p",
-            str(qmd_file),
-        ],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-    assert result.returncode in (0, 1)
-
-
-def test_main_no_qmd_files_functional(tmp_path):
-    """Functional Test: main() exits with error if no .qmd files are found."""
-    # Attempt to lint a non-existent .qmd file
-    result = subprocess.run(
-        [
-            sys.executable,
-            "-m",
-            "lintquarto",
-            "-l",
-            CORE_LINTER,
-            "-p",
-            str(tmp_path / "nofiles"),
-        ],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    # Assert that the exit code is 1 (error), and that error message is present
-    assert result.returncode == 1
-    assert "No .qmd files found" in result.stderr
-
-
-def test_decorator():
-    """Functional Test: blank line warning disabled despite decorator."""
-    # Locate the Quarto example file containing a function decorated with a
-    # runtime-checkable decorator
-    test_dir = Path(__file__).parent
-    qmd_path = test_dir / "examples" / "decorator_example.qmd"
-
-    # Run lintquarto with flake8 on the example file.
-    # Normally, flake8 would raise E302 ("expected 2 blank lines before
-    # function definition"), but lintquarto should suppress this warning
-    result = subprocess.run(
-        [sys.executable, "-m", "lintquarto", "-l", "flake8", "-p", qmd_path],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-    output = result.stdout + result.stderr
-
-    # Verify that the E302 warning does not appear in the lint output
-    assert "E302" not in output
-
-
-def test_eval_false():
-    """Functional Test: eval=false in document YAML."""
-    # Locate the Quarto example file
-    test_dir = Path(__file__).parent
-    qmd_path = test_dir / "examples" / "eval_example.qmd"
-
-    # Run lintquarto with flake8 on the example file.
-    result = subprocess.run(
-        [sys.executable, "-m", "lintquarto", "-l", "flake8", "-p", qmd_path],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-    output = result.stdout + result.stderr
-
-    # Verify that the warnings from the second chunk (eval=True) appear
-    assert "E305" in output
-    assert "F401" in output
-    assert "E402" in output
-
-    # Verify that the warnings from the first chunk do not
-    assert "501" not in output
-
-
-def test_paths_with_commas(monkeypatch):
-    """Functional Test: raises ValueError when --paths contains commas."""
-    test_args = ["prog", "-l", "pylint", "-p", "file1.qmd,dir2"]
-    monkeypatch.setattr(sys, "argv", test_args)
-    with pytest.raises(ValueError, match="contains a comma"):
-        main()
-
-
-def test_exclude_with_commas(monkeypatch):
-    """Functional Test: raises ValueError when --exclude contains commas."""
-    test_args = [
-        "prog",
-        "-l",
-        "pylint",
-        "-p",
-        "file1.qmd",
-        "-e",
-        "dir2,file2.qmd",
-    ]
-    monkeypatch.setattr(sys, "argv", test_args)
-    with pytest.raises(ValueError, match="contains a comma"):
-        main()
